Move tests to JsonDataTest
